requirements.txt
scipy

pandas
Django==1.8.3
Unipath==1.1
dj-database-url==0.3.0
python-decouple==2.3
requests==2.7.0

matlab
numpy-1.12.1+mkl
hdf5storage==0.1.14

django-tables2
sklearn

matplotlib
seaborn